Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigate-to/spv-only-sent-to-initiator.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'navigate-to'.



Harness Error (TIMEOUT), message = null

NOTRUN Test that no spv event is raised
PASS Violation report status OK.
